+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigDropdown.java
@@ -160,7 +160,7 @@ public class ConfigDropdown extends BasicOption {
     }
 
     @Override
-    public int getHeight() {
+    public float getHeight() {
         return 32;
     }
 }
