+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigInfo.java
@@ -33,7 +33,7 @@ public class ConfigInfo extends BasicOption {
     }
 
     @Override
-    public int getHeight() {
+    public float getHeight() {
         return 32;
     }
 }
