+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigSwitch.java
@@ -62,7 +62,7 @@ public class ConfigSwitch extends BasicOption {
     }
 
     @Override
-    public int getHeight() {
+    public float getHeight() {
         return 32;
     }
 }
